backend/requirements.txt
#######################################################
### Base & Server                                   ###
#######################################################
django
djangorestframework

#######################################################
### Tenant                                          ###
#######################################################
# A very new version is needed to be compatible with django 1.10.
django-polymorphic==1.0b1

#######################################################
### Database                                        ###
#######################################################
dj-database-url
